Change recipes/smelting to use a "global" ID, including mod name
Previously, smelting recipes were keyed by item ID, which of course
is not globally unique (if only..), so is unsuitable for resolving
conflicts. Now it is keyed by a new value, generated by getGlobalItemName(),
containing the mod ID (from FML GameData), a slash, and then either the
unlocalized name, or the item ID if the unlocalized name is insufficiently unique.

NOTE: Requires re-running modanalyzer.py to get this new data!